refactor: 위경도 변환관련 클래스 패키지 변경
