+++ src/test/java/com/map/gaja/client/presentation/web/WebClientControllerTest.java
@@ -11,7 +11,7 @@ import com.map.gaja.global.authentication.AuthenticationRepository;
 import com.map.gaja.global.authentication.PrincipalDetails;
 import com.map.gaja.group.application.GroupAccessVerifyService;
 import com.map.gaja.group.application.GroupService;
-import com.map.gaja.client.infrastructure.location.Geocoder;
+import com.map.gaja.client.application.geocode.Geocoder;
 import com.map.gaja.user.domain.model.Authority;
 
 import org.junit.jupiter.api.BeforeEach;
